Module aws_sdk_codestar::error

Modules

concurrent_modification_exception

See ConcurrentModificationException

invalid_next_token_exception

See InvalidNextTokenException

invalid_service_role_exception

See InvalidServiceRoleException

limit_exceeded_exception

See LimitExceededException

project_already_exists_exception

See ProjectAlreadyExistsException

project_configuration_exception

See ProjectConfigurationException

project_creation_failed_exception

See ProjectCreationFailedException

project_not_found_exception

See ProjectNotFoundException

team_member_already_associated_exception

See TeamMemberAlreadyAssociatedException

team_member_not_found_exception

See TeamMemberNotFoundException

user_profile_already_exists_exception

See UserProfileAlreadyExistsException

user_profile_not_found_exception

See UserProfileNotFoundException

validation_exception

See ValidationException

Structs

AssociateTeamMemberError

ConcurrentModificationException

Another modification is being made. That modification must complete before you can make

CreateProjectError

CreateUserProfileError

DeleteProjectError

DeleteUserProfileError

DescribeProjectError

DescribeUserProfileError

DisassociateTeamMemberError

InvalidNextTokenException

The next token is not valid.

InvalidServiceRoleException

The service role is not valid.

LimitExceededException

A resource limit has been exceeded.

ListProjectsError

ListResourcesError

ListTagsForProjectError

ListTeamMembersError

ListUserProfilesError

ProjectAlreadyExistsException

An AWS CodeStar project with the same ID already exists in this region for the AWS account.

ProjectConfigurationException

Project configuration information is required but not specified.

ProjectCreationFailedException

The project creation request was valid, but a nonspecific exception or error occurred

ProjectNotFoundException

The specified AWS CodeStar project was not found.

TagProjectError

TeamMemberAlreadyAssociatedException

The team member is already associated with a role in this project.

TeamMemberNotFoundException

The specified team member was not found.

UntagProjectError

UpdateProjectError

UpdateTeamMemberError

UpdateUserProfileError

UserProfileAlreadyExistsException

A user profile with that name already exists in this region for the AWS account. AWS

UserProfileNotFoundException

The user profile was not found.

ValidationException

The specified input is either not valid, or it could not be validated.

Enums

AssociateTeamMemberErrorKind

CreateProjectErrorKind

CreateUserProfileErrorKind

DeleteProjectErrorKind

DeleteUserProfileErrorKind

DescribeProjectErrorKind

DescribeUserProfileErrorKind

DisassociateTeamMemberErrorKind

ListProjectsErrorKind

ListResourcesErrorKind

ListTagsForProjectErrorKind

ListTeamMembersErrorKind

ListUserProfilesErrorKind

TagProjectErrorKind

UntagProjectErrorKind

UpdateProjectErrorKind

UpdateTeamMemberErrorKind

UpdateUserProfileErrorKind