Module aws_sdk_codestar::model

Modules

code

See Code

code_commit_code_destination

See CodeCommitCodeDestination

code_destination

See CodeDestination

code_source

See CodeSource

git_hub_code_destination

See GitHubCodeDestination

project_status

See ProjectStatus

project_summary

See ProjectSummary

resource

See Resource

s3_location

See S3Location

team_member

See TeamMember

toolchain

See Toolchain

toolchain_source

See ToolchainSource

user_profile_summary

See UserProfileSummary

Structs

Code

Location and destination information about the source code files provided with the

CodeCommitCodeDestination

Information about the AWS CodeCommit repository to be created in AWS CodeStar. This is where the source code files provided

CodeDestination

The repository to be created in AWS CodeStar. Valid values are AWS CodeCommit or GitHub. After AWS CodeStar provisions the new repository, the

CodeSource

The location where the source code files provided with the project request are

GitHubCodeDestination

Information about the GitHub repository to be created in AWS CodeStar. This is where the source code files provided with the

ProjectStatus

An indication of whether a project creation or deletion is failed or

ProjectSummary

Information about the metadata for a project.

Resource

Information about a resource for a project.

S3Location

The Amazon S3 location where the source code files provided with the project

TeamMember

Information about a team member in a project.

Toolchain

The toolchain template file provided with the project request. AWS CodeStar uses

ToolchainSource

The Amazon S3 location where the toolchain template file provided with the

UserProfileSummary

Information about a user's profile in AWS CodeStar.