Module aws_sdk_codestar::operation

Structs

AssociateTeamMember

Adds an IAM user to the team for an AWS CodeStar project.

CreateProject

Creates a project, including project resources. This action creates a project based on

CreateUserProfile

Creates a profile for a user that includes user preferences, such as the display name

DeleteProject

Deletes a project, including project resources. Does not delete users associated with

DeleteUserProfile

Deletes a user profile in AWS CodeStar, including all personal preference data associated with

DescribeProject

Describes a project and its resources.

DescribeUserProfile

Describes a user in AWS CodeStar and the user attributes across all projects.

DisassociateTeamMember

Removes a user from a project. Removing a user from a project also removes the IAM

ListProjects

Lists all projects in AWS CodeStar associated with your AWS account.

ListResources

Lists resources associated with a project in AWS CodeStar.

ListTagsForProject

Gets the tags for a project.

ListTeamMembers

Lists all team members associated with a project.

ListUserProfiles

Lists all the user profiles configured for your AWS account in AWS CodeStar.

TagProject

Adds tags to a project.

UntagProject

Removes tags from a project.

UpdateProject

Updates a project in AWS CodeStar.

UpdateTeamMember

Updates a team member's attributes in an AWS CodeStar project. For example, you can change a

UpdateUserProfile

Updates a user's profile in AWS CodeStar. The user profile is not project-specific.